requirements.txt
AWSIoTPythonSDK==1.4.9
paho-mqtt==1.6.1
